+++ contactos/models.py
@@ -25,7 +25,7 @@ class Contact(models.Model):
     value = models.CharField(max_length=512)
     is_bounced = models.BooleanField(default=False)
     owner = models.ForeignKey(User, related_name="contacts", null=True)
-    writeitinstance = models.ForeignKey('nuntium.WriteItInstance', related_name="contacts", null=True)
+    writeitinstance = models.ForeignKey('instance.WriteItInstance', related_name="contacts", null=True)
     popit_identifier = models.CharField(max_length=512, null=True)
     enabled = models.BooleanField(default=True)
